Added offline suport (#4)
* Created an adapter interface for operations

This also moves all the code that talks directly to gitignore.io into a
separate adapter. This makes it easier to add other implementations that
might read directly from the file system or other sources.

* Added a filesystem based adapter

This will serve as a caching layer but that part of things isn't
implemented yet.

* Removed gitignore.io headers

This makes the output from the filesystem adapter the same as the HTTP
adapter for consistency. This is also something I do manually normally
so automating it seems a good idea.

* Added specialized adapter types

This allows adapters to specify if they are a source of data or a sink
for data. The main use case here is to allow the HTTP adapter to
generate content that is cached by the filesystem adapter.

* Don't include debug information in final build

* Run tests in parallel when possible

* Implemented missing test

* Added an update feature for caching options

* Added Git based adapter

This replaces both the filesystem and HTTP adapters as it supports
remote updates and local caching.