fix(git): Fix update error for existing repo
If the the repository already exists then an exceptionw as being thrown.
This was because we were always assuming that we needed to clone a new
repository but this would break if the directory already exists as a git
repository. In that case, we need should be doing a `git pull` instead.